split up isExpanded
